+++ businesses.go
@@ -4,6 +4,7 @@ package paddle
 
 import (
 	"context"
+
 	paddleerr "github.com/PaddleHQ/paddle-go-sdk/pkg/paddleerr"
 )