+++ events.go
@@ -5,6 +5,7 @@ package paddle
 import (
 	"context"
 	"encoding/json"
+
 	paddlenotification "github.com/PaddleHQ/paddle-go-sdk/pkg/paddlenotification"
 )